// src/api/service/model/recipeDto.ts
/**
 * Things Keeper
 * Will save your notes.
 *
 * OpenAPI spec version: 0.1.0
 * Contact: vrahlinskiy@cinimex.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageDto } from './imageDto';
import { IngredientDto } from './ingredientDto';
import { StepDto } from './stepDto';
import { Tag } from './tag';


export interface RecipeDto { 
    category?: string;
    created?: Date;
    description?: string;
    id?: string;
    image?: ImageDto;
    ingredients?: Array<IngredientDto>;
    link?: string;
    steps?: Array<StepDto>;
    tags?: Array<Tag>;
    title?: string;
}